Fixed ping waypoints sending with .0 Added key spawn announce Added option to disable BlockOverlay when using Etherwarp

+++ src/main/kotlin/me/odinmain/commands/impl/WaypointCommand.kt
@@ -30,7 +30,7 @@ val waypointCommand = commodore("waypoint", "odinwaypoint") {
         runs {
             partyMessage(PlayerUtils.getPositionString())
         }
-        runs { x: Double, y: Double, z: Double ->
+        runs { x: Int, y: Int, z: Int ->
             partyMessage("x: $x y: $y, z: $z")
         }
     }
